+++ stroom-auth-service-api/src/main/java/stroom/auth/service/api/model/PasswordValidationRequest.java
@@ -13,17 +13,16 @@
 
 package stroom.auth.service.api.model;
 
+import java.util.Objects;
 import com.google.gson.annotations.SerializedName;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
-
-import java.util.Objects;
 
 /**
  * A request to validate a user&#39;s password.
  */
 @ApiModel(description = "A request to validate a user's password.")
-@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaClientCodegen", date = "2019-11-11T12:01:59.457Z")
+@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaClientCodegen", date = "2019-11-26T10:17:08.903Z")
 public class PasswordValidationRequest {
   @SerializedName("oldPassword")
   private String oldPassword = null;