+++ stroom-auth-svc/src/main/java/stroom/auth/TokenBuilder.java
@@ -40,9 +40,15 @@ public class TokenBuilder {
     private Optional<String> state = Optional.empty();
     private PrivateKey privateVerificationKey;
     private String authSessionId;
+    private String clientId;
 
     public TokenBuilder subject(String subject) {
         this.subject = subject;
+        return this;
+    }
+
+    public TokenBuilder clientId(String clientId) {
+        this.clientId = clientId;
         return this;
     }
 
@@ -116,6 +122,7 @@ public class TokenBuilder {
         claims.setExpirationTime(NumericDate.fromSeconds(expiryDate.getEpochSecond()));
         claims.setSubject(subject);
         claims.setIssuer(issuer);
+        claims.setAudience(clientId);
         claims.setStringClaim("sid", authSessionId);
         nonce.ifPresent(nonce -> claims.setClaim("nonce", nonce));
         state.ifPresent(state -> claims.setClaim("state", state));